Added Sierpinski's Triangle and Particle Art
These were made using the P5.js editor after watching some videos from The Coding Train on YouTube. The Sierpinski's Triangle one was created for a lesson I taught for the ASU Computer Science club when I was a leader in the group. The Particle Acceleration Art was created by following The Coding Train's video on boids. When all particles start at a different location and velocity, but the same acceleration and looping side walls, they create a beautiful pattern.